+++ src/Formatter.php
@@ -20,24 +20,47 @@ class Formatter
         '\bMacKlin'    => 'Macklin',
         '\bMacKmin'    => 'Mackmin',
         '\bMacQuarie'  => 'Macquarie',
+        '\bMacOmber'   => 'Macomber',
+        '\bMacIn'      => 'Macin',
+        '\bMacKintosh' => 'Mackintosh',
+        '\bMacKen'     => 'Macken',
+        '\bMacHen'     => 'Machen',
+        '\bMacisaac'   => 'MacIsaac',
+        '\bMacHiel'    => 'Machiel',
+        '\bMacIol'     => 'Maciol',
+        '\bMacKell'    => 'Mackell',
+        '\bMacKlem'    => 'Macklem',
+        '\bMacKrell'   => 'Mackrell',
+        '\bMacLin'     => 'Maclin',
+        '\bMacKey'     => 'Mackey',
+        '\bMacKley'    => 'Mackley',
+        '\bMacHell'    => 'Machell',
+        '\bMacHon'     => 'Machon',
     ];
 
     // General replacements.
     private const REPLACEMENTS = [
         '\bAl(?=\s+\w)'         => 'al',        // al Arabic or forename Al.
-        '\b(Bin|Binti|Binte)\b' => 'bin',       // bin, binti, binte Arabic
         '\bAp\b'                => 'ap',        // ap Welsh.
-        '\bBen(?=\s+\w)'        => 'ben',       // ben Hebrew or forename Ben.
+        '\b(Bin|Binti|Binte)\b' => 'bin',       // bin, binti, binte Arabic.
         '\bDell([ae])\b'        => 'dell\1',    // della and delle Italian.
-        '\bD([aeiou])\b'        => 'd\1',       // da, de, di Italian; du French; do Brasil
-        '\bD([ao]s)\b'          => 'd\1',       // das, dos Brasileiros
+        '\bD([aeiou])\b'        => 'd\1',       // da, de, di Italian; du French; do Brasil.
+        '\bD([ao]s)\b'          => 'd\1',       // das, dos Brasileiros.
         '\bDe([lrn])\b'         => 'de\1',      // del Italian; der/den Dutch/Flemish.
-        '\bEl\b'                => 'el',        // el Greek or El Spanish.
-        '\bLa\b'                => 'la',        // la French or La Spanish.
         '\bL([eo])\b'           => 'l\1',       // lo Italian; le French.
-        '\bTe([rn])'            => 'te\1',      // ten, ter Dutch/Flemish
+        '\bTe([rn])'            => 'te\1',      // ten, ter Dutch/Flemish.
         '\bVan(?=\s+\w)'        => 'van',       // van German or forename Van.
-        '\bVon\b'               => 'von',       // von Dutch/Flemish
+        '\bVon\b'               => 'von',       // von Dutch/Flemish.
+    ];
+
+    private const SPANISH = [
+        '\bEl\b' => 'el',        // el Greek or El Spanish.
+        '\bLa\b' => 'la',        // la French or La Spanish.
+    ];
+
+    const HEBREW = [
+        '\bBen(?=\s+\w)' => 'ben', // ben Hebrew or forename Ben.
+        '\bBat(?=\s+\w)' => 'bat', // bat Hebrew or forename Bat.
     ];
 
     // Spanish conjunctions.
@@ -46,79 +69,153 @@ class Formatter
     // Roman letters regexp.
     private const ROMAN_REGEX = '\b((?:[Xx]{1,3}|[Xx][Ll]|[Ll][Xx]{0,3})?(?:[Ii]{1,3}|[Ii][VvXx]|[Vv][Ii]{0,3})?)\b';
 
+    // Post nominal values.
+    private const POST_NOMINALS = [
+        'VC', 'GC', 'KG', 'LG', 'KT', 'LT', 'KP', 'GCB', 'OM', 'GCSI', 'GCMG', 'GCIE', 'GCVO',
+        'GBE', 'CH', 'KCB', 'DCB', 'KCSI', 'KCMG', 'DCMG', 'KCIE', 'KCVO', 'DCVO', 'KBE', 'DBE',
+        'CB', 'CSI', 'CMG', 'CIE', 'CVO', 'CBE', 'DSO', 'LVO', 'OBE', 'ISO', 'MVO', 'MBEIOM', 'CGC',
+        'RRC', 'DSC', 'MC', 'DFC', 'AFC', 'ARRC', 'OBI', 'DCM', 'CGM', 'GM', 'IDSM', 'DSM', 'MM',
+        'DFM', 'AFM', 'SGM', 'IOMCPM', 'QGM', 'RVM', 'BEM', 'QPM', 'QFSM', 'QAM', 'CPM', 'MSM',
+        'ERD', 'VD', 'TD', 'UD', 'ED', 'RD', 'VRD', 'AEPC', 'ADC', 'QHP', 'QHS', 'QHDS', 'QHNS',
+        'QHC', 'SCJ', 'J', 'LJ', 'QS', 'SL', 'QC', 'KC', 'JP', 'DL', 'MP', 'MSP', 'MSYP', 'AM',
+        'MLA', 'MEP', 'DBEnv', 'DConstMgt', 'DREst', 'EdD', 'DPhil', 'PhD', 'DLitt', 'DSocSci',
+        'MD', 'EngD', 'DD', 'LLD', 'DProf', 'MA', 'MArch', 'MAnth', 'MSc', 'MMORSE', 'MMath',
+        'MMathStat', 'MPharm', 'MPhil', 'MSc', 'MSci', 'MSt', 'MRes', 'MEng', 'MChem', 'MBiochem',
+        'MSocSc', 'MMus', 'LLM', 'BCL', 'MPhys', 'MComp', 'MAcc', 'MFin', 'MBA', 'MPAMEd', 'MEP',
+        'MEnt', 'MCGI', 'MGeol', 'MLitt', 'MEarthSc', 'MClinRes', 'BA', 'BSc', 'LLB', 'BEng',
+        'MBChB', 'FdAFdSc', 'FdEng', 'PgDip', 'PgD', 'PgCert', 'PgC', 'PgCLTHE', 'AUH', 'AKC',
+        'AUS', 'HNC', 'HNCert', 'HND', 'HNDip', 'DipHE', 'Dip', 'OND', 'CertHE', 'ACSM', 'MCSM',
+        'DIC', 'AICSM', 'ARSM', 'ARCS', 'LLB', 'LLM', 'BCL', 'MJur', 'DPhil', 'PhD', 'LLD', 'DipLP',
+        'FCILEx', 'GCILEx', 'ACILEx', 'CQSW', 'DipSW', 'BSW', 'MSW', 'FCILT', 'CMILT', 'MILT',
+        'CPLCTP', 'CML', 'PLS', 'CTL', 'DLP', 'PLog', 'EJLog', 'ESLog', 'EMLog', 'JrLog', 'Log',
+        'SrLog', 'BArch', 'MArch', 'ARBRIBA', 'RIAS', 'RIAI', 'RSAW', 'MB', 'BM', 'BS', 'BCh',
+        'BChir', 'MRCS', 'FRCS', 'MS', 'MCh.', 'MRCP', 'FRCP', 'MRCPCHFRCPCH', 'MRCPath', 'MFPM',
+        'FFPM', 'BDS', 'MRCPsych', 'FRCPsych', 'MRCOG', 'FRCOG', 'MCEM', 'FCEM', 'FRCAFFPMRCA',
+        'MRCGP', 'FRCGP', 'BSc', 'MScChiro', 'MChiro', 'MSc', 'DC', 'LFHOM', 'MFHOM', 'FFHOM',
+        'FADO', 'FBDOFCOptom', 'MCOptom', 'MOst', 'DPT', 'MCSP', 'FCSP.', 'SROT', 'MSCR', 'FSCR.',
+        'CPhT', 'RN', 'VN', 'RVN', 'BVScBVetMed', 'VetMB', 'BVM&S', 'MRCVS', 'FRCVS', 'FAWM',
+        'PGCAP', 'PGCHE', 'PGCE', 'PGDE', 'BEd', 'NPQH', 'QTSCSci', 'CSciTeach', 'RSci', 'RSciTech',
+        'CEng', 'IEng', 'EngTech', 'ICTTech', 'DEM', 'MM', 'CMarEngCMarSci', 'CMarTech', 'IMarEng',
+        'MarEngTech', 'RGN', 'SRN', 'RMN', 'RSCN', 'SEN', 'EN', 'RNMH', 'RN', 'RM', 'RN1RNA', 'RN2',
+        'RN3', 'RNMH', 'RN4', 'RN5', 'RNLD', 'RN6', 'RN8', 'RNC', 'RN7', 'RN9', 'RHV', 'RSN', 'ROH',
+        'RFHN', 'SPANSPMH', 'SPCN', 'SPLD', 'SPHP', 'SCHM', 'SCLD', 'SPCC', 'SPDN', 'V100', 'V200',
+        'V300', 'LPE', 'MS'
+    ];
+
     // Default options.
     private static $options = [
-        'lazy'    => true,
-        'irish'   => true,
-        'spanish' => true,
+        'lazy'        => true,
+        'irish'       => true,
+        'spanish'     => false,
+        'roman'       => true,
+        'hebrew'      => true,
+        'postnominal' => true,
     ];
+
+    /**
+     * Formatter constructor.
+     *
+     * @param array $options
+     */
+    public function __construct ($options = [])
+    {
+        $this->setOptions($options);
+    }
+
+    /**
+     * Global options setter.
+     *
+     * @param array $options
+     */
+    public static function setOptions ($options)
+    {
+        self::$options = array_merge(self::$options, $options);
+    }
 
     /**
      * Main function for NameCase.
      *
-     * @param string $string
-     * @param array  $options
+     * @param string $name
+     * @param array $options
      *
      * @return string
      */
-    public static function nameCase(string $string = '', array $options = []): string
+    public static function nameCase ($name = '', array $options = []): string
     {
-        if ($string == '') return $string;
+        if ($name == '') return $name;
 
         self::$options = array_merge(self::$options, $options);
 
         // Do not do anything if string is mixed and lazy option is true.
-        if (self::$options['lazy'] && self::skipMixed($string)) return $string;
+        if (self::$options['lazy'] && self::skipMixed($name)) return $name;
 
         // Capitalize
-        $string = self::capitalize($string);
-        $string = self::updateIrish($string);
+        $name = self::capitalize($name);
+        $name = self::updateIrish($name);
 
-        // Fixes for "son (daughter) of" etc
-        foreach (self::REPLACEMENTS as $pattern => $replacement) {
-            $string = mb_ereg_replace($pattern, $replacement, $string);
+        // General fixes
+        $replacements = self::REPLACEMENTS;
+        if ( ! self::$options['spanish']) {
+            $replacements = array_merge($replacements, self::SPANISH);
         }
 
-        $string = self::updateRoman($string);
-        $string = self::fixConjunction($string);
+        if (self::$options['hebrew']) {
+            $replacements = array_merge($replacements, self::HEBREW);
+        }
 
-        return $string;
+        foreach ($replacements as $pattern => $replacement) {
+            $name = mb_ereg_replace($pattern, $replacement, $name);
+        }
+
+        if (self::$options['roman']) {
+            $name = self::updateRoman($name);
+        }
+
+        if (self::$options['spanish']) {
+            $name = self::fixConjunction($name);
+        }
+
+        if (self::$options['postnominal']) {
+            $name = self::fixPostNominal($name);
+        }
+
+        return $name;
     }
 
     /**
      * Capitalize first letters.
      *
-     * @param string $string
+     * @param string $name
      *
      * @return string
      */
-    private static function capitalize(string $string): string
+    private static function capitalize (string $name): string
     {
-        $string = mb_strtolower($string);
+        $name = mb_strtolower($name);
 
-        $string = mb_ereg_replace_callback('\b\w', function ($matches) {
+        $name = mb_ereg_replace_callback('\b\w', function ($matches) {
             return mb_strtoupper($matches[0]);
-        }, $string);
+        }, $name);
 
         // Lowercase 's
-        $string = mb_ereg_replace_callback('\'\w\b', function ($matches) {
+        $name = mb_ereg_replace_callback('\'\w\b', function ($matches) {
             return mb_strtolower($matches[0]);
-        }, $string);
+        }, $name);
 
-        return $string;
+        return $name;
     }
 
     /**
      * Skip if string is mixed case.
      *
-     * @param string $string
+     * @param string $name
      *
      * @return bool
      */
-    private static function skipMixed(string $string): bool
+    private static function skipMixed (string $name): bool
     {
-        $firstLetterLower = $string[0] == mb_strtolower($string[0]);
-        $allLowerOrUpper  = (mb_strtolower($string) == $string || mb_strtoupper($string) == $string);
+        $firstLetterLower = $name[0] == mb_strtolower($name[0]);
+        $allLowerOrUpper = (mb_strtolower($name) == $name || mb_strtoupper($name) == $name);
 
         return ! ($firstLetterLower || $allLowerOrUpper);
     }
@@ -126,71 +223,85 @@ class Formatter
     /**
      * Update for Irish names.
      *
-     * @param string $string
+     * @param string $name
      *
      * @return string
      */
-    private static function updateIrish(string $string): string
+    private static function updateIrish (string $name): string
     {
-        if ( ! self::$options['irish']) return $string;
+        if ( ! self::$options['irish']) return $name;
 
-        if (mb_ereg_match('.*?\bMac[A-Za-z]{2,}[^aciozj]\b', $string) || mb_ereg_match('.*?\bMc', $string)) {
-            $string = self::updateMac($string);
+        if (
+            mb_ereg_match('.*?\bMac[A-Za-z]{2,}[^aciozj]\b', $name) ||
+            mb_ereg_match('.*?\bMc', $name)
+        ) {
+            $name = self::updateMac($name);
         }
 
-        return mb_ereg_replace('Macmurdo', 'MacMurdo', $string);
-    }
-
-    /**
-     * Fix Spanish conjunctions.
-     *
-     * @param string $string
-     *
-     * @return string
-     */
-    private static function fixConjunction(string $string): string
-    {
-        if ( ! self::$options['spanish']) return $string;
-
-        foreach (self::CONJUNCTIONS as $conjunction) {
-            $string = mb_ereg_replace('\b' . $conjunction . '\b', mb_strtolower($conjunction), $string);
-        }
-
-        return $string;
+        return mb_ereg_replace('Macmurdo', 'MacMurdo', $name);
     }
 
     /**
      * Fix roman numeral names.
      *
-     * @param string $string
+     * @param string $name
      *
      * @return string
      */
-    private static function updateRoman(string $string): string
+    private static function updateRoman (string $name): string
     {
         return mb_ereg_replace_callback(self::ROMAN_REGEX, function ($matches) {
             return mb_strtoupper($matches[0]);
-        }, $string);
+        }, $name);
     }
 
     /**
      * Updates irish Mac & Mc.
      *
-     * @param string $string
+     * @param string $name
      *
      * @return string
      */
-    private static function updateMac(string $string): string
+    private static function updateMac (string $name): string
     {
-        $string = mb_ereg_replace_callback('\b(Ma?c)([A-Za-z]+)', function ($matches) {
+        $name = mb_ereg_replace_callback('\b(Ma?c)([A-Za-z]+)', function ($matches) {
             return $matches[1] . mb_strtoupper(mb_substr($matches[2], 0, 1)) . mb_substr($matches[2], 1);
-        }, $string);
+        }, $name);
 
         // Now fix "Mac" exceptions
         foreach (self::EXCEPTIONS as $pattern => $replacement) {
-            $string = mb_ereg_replace($pattern, $replacement, $string);
+            $name = mb_ereg_replace($pattern, $replacement, $name);
         }
 
-        return $string;
+        return $name;
+    }
+
+    /**
+     * Fix Spanish conjunctions.
+     *
+     * @param string $name
+     *
+     * @return string
+     */
+    private static function fixConjunction (string $name): string
+    {
+        foreach (self::CONJUNCTIONS as $conjunction) {
+            $name = mb_ereg_replace('\b' . $conjunction . '\b', mb_strtolower($conjunction), $name);
+        }
+        return $name;
+    }
+
+    /**
+     * Fix post-nominal letter cases.
+     *
+     * @param string $name
+     * @return string
+     */
+    private static function fixPostNominal (string $name): string
+    {
+        foreach (self::POST_NOMINALS as $postnominal) {
+            $name = mb_ereg_replace('\b' . $postnominal . '$', $postnominal, $name, 'ix');
+        }
+        return $name;
     }
 }